requirements.txt
# ── Web UI ───────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LangChain core ───────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# ── OpenAI (optional – only needed if USE_OPENAI=true) ───────────────────────────
langchain-openai>=0.1.0
openai>=1.30.0

# ── Embeddings (HuggingFace / sentence-transformers) ─────────────────────────────
sentence-transformers>=2.7.0
transformers>=4.40.0,<4.50.0
torch>=2.1.0
torchvision>=0.16.0

# ── Vector store ─────────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── PDF parsing ──────────────────────────────────────────────────────────────────
PyMuPDF>=1.24.0

# ── Utilities ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tiktoken>=0.7.0
numpy>=1.26.0